% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnadc_grafico.R
\name{pnadc_graph}
\alias{pnadc_graph}
\title{pnadc_graph}
\usage{
pnadc_graph(
  variable,
  filter,
  year,
  quartile,
  calculation,
  classifier,
  path = F,
  export = F,
  type = 1
)
}
\arguments{
\item{variable}{Variable of interest that will be used to calculate the Total. It can be a vector of variables.}

\item{filter}{Variable that defines the aggregation level of the variable of interest. It can contain more than one level of aggregation.}

\item{year}{The year you want to review. Must be a number between 2012 and the current year. It does not accept a vector of years.}

\item{quartile}{The quartile of the year you want to analyze. Must be a number between 1 and 4. It does not accept an array of quartiles.}

\item{calculation}{calculation what you want to do. Must be mean, total or percentage and come between "". It does not accept a list of calculations.}

\item{classifier}{One of the variables that were used in the 'variable' or 'filter' parameter that will serve as a classifier filter in the legend and graph. It must come between "".}

\item{path}{Path to the directory where the Design was created through the function "pnadc_download". It only accepts logical values True or False.}

\item{export}{Export the chart to "pdf" and "png" formats. If it has not been filled, it returns a ggplot2 graph in R space. It must be a string and be enclosed in " ".}

\item{type}{Type of chart you want to generate. 1 for Bar Graph and 2 for Dot Graph. Does not accept an array of types.}
}
\value{
ggplot2 graphic
}
\description{
Creates PNADc bar or dot graphics
}
\examples{
pnadc_graph(~V403312, ~UF, 2019, 1, calculation = "mean", classifier = "V403312")
}
