% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnadc_media.R
\name{pnadc_mean}
\alias{pnadc_mean}
\title{pnadc_mean}
\usage{
pnadc_mean(variable, filter, year, quartile, path = F, export = F)
}
\arguments{
\item{variable}{Variable of interest that will be used to calculate the mean.}

\item{filter}{Variable that defines the aggregation level of the variable of interest. It can contain more than one level of aggregation.}

\item{year}{The year you want to review. Must be a number between 2012 and the current year. It does not accept a vector of years.}

\item{quartile}{The quartile of the year you want to analyze. Must be a number between 1 and 4. It does not accept a vector of quartiles.}

\item{path}{Location/directory where you want to save the final table. If you have already used the "pnadc_download" function to create the design in a specific path, just use the logical value True.}

\item{export}{Export the table to "html", "pdf", "png" and "csv" formats. If you want to return a data frame in R space, use "df". If it has not been filled, it returns a gt_table in R space. It must be a string and be enclosed in " ".}
}
\value{
gt table.
}
\description{
Create PNADc mean tables
}
\examples{
pnadc_mean(~V403312, ~UF+V2010, 2019, 1)
}
