% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnadc_porcent.R
\name{pnadc_porcent}
\alias{pnadc_porcent}
\title{Porcentagem PNADc}
\usage{
pnadc_porcent(variable, filter, year, quartile, path = F, export = F)
}
\arguments{
\item{variable}{Variable of interest that will be used to calculate the percentage.}

\item{filter}{Variable that defines the aggregation level of the variable of interest. It can contain more than one level of aggregation.}

\item{year}{The year you want to review. Must be a number between 2012 and the current year. It does not accept a vector of years.}

\item{quartile}{The quartile of the year you want to analyze. Must be a number between 1 and 4. It does not accept a vector of quartiles.}

\item{path}{Path to the directory where the Design was created using the "low_pnadc" function. It only accepts logical values True or False.}

\item{export}{Export the table to "html", "pdf", "png" and "rtf" formats. If you want to return a data frame in R space, use "df". If it has not been filled, it returns a gt table in R space. It must be a string and be enclosed in " ".}
}
\value{
gt table.
}
\description{
Create PNADc percent tables
}
\examples{
pnadc_porcent(~V403312, ~UF+V2007, 2019, 1)
}
