% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnadc_faixas.R
\name{pnadc_tracks}
\alias{pnadc_tracks}
\title{pnadc_tracks}
\usage{
pnadc_tracks(
  variable,
  filter,
  year,
  quartile,
  calculation,
  group,
  cluster,
  path = F,
  export = F
)
}
\arguments{
\item{variable}{Variable of interest that will be used to calculate the data. It must be a formula, that is, have ~ in front of the variable.}

\item{filter}{Variable that defines the aggregation level of the variable of interest. It can contain more than one level of aggregation. It must be a formula, i.e. have ~ in front of the variable.}

\item{year}{The year you want to review. Must be a number between 2012 and the current year. It does not accept a vector of years.}

\item{quartile}{The quartile of the year you want to analyze. Must be a number between 1 and 4. It does not accept a vector of quartiles.}

\item{calculation}{calculation what you want to do. Must be mean, total or percentage and come between "". It does not accept a list of calculations.}

\item{group}{Variable used to group a set of data. It must be one of the variables listed in the "filter". Must be a formula, that is, have ~ in front of the variable.}

\item{cluster}{How the grouping strip will be made. In the numerical case it must be a vector calculation the minimum, maximum and size of the range. In the categorical case it must be a list calculation the old names and the new name that will replace them.}

\item{path}{Path to the directory where the Design was created through the function "pnadc_download". It only accepts logical values True or False.}

\item{export}{Export the table to "html", "pdf", "png" and "rtf" formats. If you want to return a data frame in R space, use "df". If it has not been filled, it returns a gt table in R space. It must be a string and be enclosed in " ".}
}
\value{
gt table
}
\description{
Regroups a numeric or non-numeric variable from the PNADc
}
\examples{
#For characters
pnadc_tracks(~V403312, ~V2010, 2019, 1, calculation = "mean", group = ~V2010,
cluster = list('Negra' = c('Preta', 'Parda'), 'Outros' = c('Amarela', 'Indigena')))
#For numbers
pnadc_tracks(~V403312, ~V2009, 2019, 1, calculation = "mean",
group = ~V2009, cluster = c(0, 100, 5))

}
