% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_pnadc.R
\name{pnadc_download}
\alias{pnadc_download}
\title{pnadc_download}
\usage{
pnadc_download(year, quartile, path = NULL)
}
\arguments{
\item{year}{The year you want to download. Must be a number between 2012 and the current year. It does not accept a vector of years.}

\item{quartile}{The quartile of the year you want to download. Must be a number between 1 and 4. It does not accept a vector of quartiles.}

\item{path}{Path of the local directory where the PNAD files are or where you want to save the data. If it does not exist, a new directory will be created.}
}
\description{
download the PNADc version and create a design directly on the computer.
}
\examples{
pnadc_download(2019, 1)
}
